requirements.txt
acres==0.2.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
declxml==0.9.1
deepdiff==8.2.0
defusedxml==0.7.1
distro==1.9.0
Django==5.1.7
docx==0.2.4
docx2txt==0.8
easyocr==1.7.2
entrypoints==0.4
etelemetry==0.3.1
faiss-cpu==1.10.0
Faker==35.2.0
fastapi==0.115.9
favicon==0.7.0
filelock==3.17.0
fitz==0.0.1.dev2
Flask==3.1.0
fonttools==4.55.8
fpdf==1.7.2
fpdf2==2.8.2
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
htbuilder==0.9.0
html2text==2024.2.26
html2txt==0.6.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain-core==0.3.49
langchain-ollama==0.3.0
langsmith==0.3.19
lazy_loader==0.4
Levenshtein==0.26.1
looseversion==1.3.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mistune==0.8.4
mpmath==1.3.0
narwhals==1.24.2
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nipype==1.9.2
numpy==2.2.2
ollama==0.4.7
openai==1.69.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orderly-set==5.3.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
plotly==6.0.0
prometheus_client==0.21.1
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyarrow==19.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
pymdown-extensions==10.14.3
PyMuPDF==1.25.2
pyparsing==3.2.1
pypdf==5.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
pyxnat==1.6.2
PyYAML==6.0.2
RapidFuzz==3.12.1
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.3
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==4.0.1
setuptools==75.8.0
shapely==2.0.7
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.45.3
streamlit==1.41.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
textdistance==4.6.3
threadpoolctl==3.6.0
tifffile==2025.1.10
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.50.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
zstandard==0.23.0
